+++ R/tm_t_abnormality.R
@@ -289,11 +289,6 @@ srv_t_abnormality <- function(input,
     adsl_vars <- unique(c("USUBJID", "STUDYID", arm_var, id_var))
     anl_vars <- unique(c("USUBJID", "STUDYID", by_vars, grade, baseline))
 
-    labels <- c(
-      datasets$get_column_labels("ADSL", adsl_vars),
-      datasets$get_column_labels(dataname, anl_vars)
-    )
-
     #nolint start
     chunks_push(bquote({
       ADSL_S <- ADSL_FILTERED[, .(adsl_vars)]
@@ -311,10 +306,12 @@ srv_t_abnormality <- function(input,
         call(
           "%>%",
           as.name("ANL_MERGED"),
-          as.call(c(
-            quote(rtables::var_relabel),
-            labels[!duplicated(names(labels))]
-          ))
+          teal.devel::get_relabel_call(
+            labels = c(
+              datasets$get_column_labels("ADSL", adsl_vars),
+              datasets$get_column_labels(dataname, anl_vars)
+            )
+          )
         )
       )
     )