+++ R/tm_t_events.R
@@ -201,15 +201,12 @@ srv_t_events_byterm <- function(input, output, session, datasets, dataname, even
               )
             ))
           ),
-          as.call(c(
-            quote(rtables::var_relabel), {
-              labels <- c(
-                datasets$get_column_labels("ADSL", adsl_vars),
-                datasets$get_column_labels(dataname, anl_vars)
-              )
-              labels[!duplicated(labels)]
-            }
-          ))
+          teal.devel::get_relabel_call(
+            labels = c(
+              datasets$get_column_labels("ADSL", adsl_vars),
+              datasets$get_column_labels(dataname, anl_vars)
+            )
+          )
         )
       )
     )