+++ R/tm_t_events_by_grade.R
@@ -240,15 +240,12 @@ srv_t_events_by_grade <- function(input, output, session, datasets, dataname) {
               )
             ))
           ),
-          as.call(c(
-            quote(rtables::var_relabel), {
-              labels <- c(
-                datasets$get_column_labels("ADSL", adsl_vars),
-                datasets$get_column_labels(dataname, anl_vars)
-              )
-              labels[!duplicated(labels)]
-            }
-          ))
+          teal.devel::get_relabel_call(
+            labels = c(
+              datasets$get_column_labels("ADSL", adsl_vars),
+              datasets$get_column_labels(dataname, anl_vars)
+            )
+          )
         )
       )
     )